include/utime: un-inline several functions to reduce header dependencies
