+++ src/include/utime.cc
@@ -14,6 +14,10 @@
 
 #include "utime.h"
 #include "common/Formatter.h"
+#include "common/strtol.h"
+#include "include/timegm.h"
+
+#include <errno.h>
 
 void utime_t::dump(ceph::Formatter *f) const
 {
@@ -28,4 +32,111 @@ void utime_t::generate_test_instances(std::list<utime_t*>& o)
   o.back()->tv.tv_sec = static_cast<__u32>((1L << 32) - 1);
   o.push_back(new utime_t());
   o.back()->tv.tv_nsec = static_cast<__u32>((1L << 32) - 1);
+}
+
+int utime_t::parse_date(const std::string& date, uint64_t *epoch, uint64_t *nsec,
+                        std::string *out_date,
+			std::string *out_time) {
+  struct tm tm;
+  memset(&tm, 0, sizeof(tm));
+
+  if (nsec)
+    *nsec = 0;
+
+  const char *p = strptime(date.c_str(), "%Y-%m-%d", &tm);
+  if (p) {
+    if (*p == ' ' || *p == 'T') {
+      p++;
+      // strptime doesn't understand fractional/decimal seconds, and
+      // it also only takes format chars or literals, so we have to
+      // get creative.
+      char fmt[32] = {0};
+      strncpy(fmt, p, sizeof(fmt) - 1);
+      fmt[0] = '%';
+      fmt[1] = 'H';
+      fmt[2] = ':';
+      fmt[3] = '%';
+      fmt[4] = 'M';
+      fmt[6] = '%';
+      fmt[7] = 'S';
+      const char *subsec = 0;
+      char *q = fmt + 8;
+      if (*q == '.') {
+	++q;
+	subsec = p + 9;
+	q = fmt + 9;
+	while (*q && isdigit(*q)) {
+	  ++q;
+	}
+      }
+      // look for tz...
+      if (*q == '-' || *q == '+') {
+	*q = '%';
+	*(q+1) = 'z';
+	*(q+2) = 0;
+      }
+      p = strptime(p, fmt, &tm);
+      if (!p) {
+	return -EINVAL;
+      }
+      if (nsec && subsec) {
+	unsigned i;
+	char buf[10]; /* 9 digit + null termination */
+	for (i = 0; (i < sizeof(buf) - 1) && isdigit(*subsec); ++i, ++subsec) {
+	  buf[i] = *subsec;
+	}
+	for (; i < sizeof(buf) - 1; ++i) {
+	  buf[i] = '0';
+	}
+	buf[i] = '\0';
+	std::string err;
+	*nsec = (uint64_t)strict_strtol(buf, 10, &err);
+	if (!err.empty()) {
+	  return -EINVAL;
+	}
+      }
+    }
+  } else {
+    int sec, usec;
+    int r = sscanf(date.c_str(), "%d.%d", &sec, &usec);
+    if (r != 2) {
+      return -EINVAL;
+    }
+
+    time_t tt = sec;
+    gmtime_r(&tt, &tm);
+
+    if (nsec) {
+      *nsec = (uint64_t)usec * 1000;
+    }
+  }
+
+  #ifndef _WIN32
+  // apply the tm_gmtoff manually below, since none of mktime,
+  // gmtime, and localtime seem to do it.  zero it out here just in
+  // case some other libc *does* apply it.  :(
+  auto gmtoff = tm.tm_gmtoff;
+  tm.tm_gmtoff = 0;
+  #else
+  auto gmtoff = _timezone;
+  #endif /* _WIN32 */
+
+  time_t t = internal_timegm(&tm);
+  if (epoch)
+    *epoch = (uint64_t)t;
+
+  *epoch -= gmtoff;
+
+  if (out_date) {
+    char buf[32];
+    strftime(buf, sizeof(buf), "%Y-%m-%d", &tm);
+    *out_date = buf;
+  }
+  if (out_time) {
+    char buf[32];
+    strftime(buf, sizeof(buf), "%H:%M:%S", &tm);
+    *out_time = buf;
+  }
+
+  return 0;
 }
