+++ src/tools/ceph-dencoder/CMakeLists.txt
@@ -11,7 +11,6 @@ endif()
 set(dencoder_srcs
   ceph_dencoder.cc
   ../../include/uuid.cc
-  ../../include/utime.cc
   $<TARGET_OBJECTS:common_texttable_obj>)
 add_executable(ceph-dencoder ${dencoder_srcs})
 set_target_properties(ceph-dencoder PROPERTIES
@@ -39,7 +38,6 @@ endfunction()
 
 add_denc_mod(denc-mod-common
   ../../include/uuid.cc
-  ../../include/utime.cc
   common_types.cc)
 target_link_libraries(denc-mod-common
   ceph-common
